[IMPROVEMENT](#780) Refactoring comment related code for code-editor [DESC.] Refactoring comment related code for code-editor into a new content assist plugin
