+++ apps/ide/src/plugins/plugin-settings-guest.json
@@ -2,6 +2,7 @@
     "plugins" : [
         "plugins/webida.editor.text-editor",
         "plugins/webida.editor.code-editor",
+        "plugins/webida.editor.code-editor.content-assist.comments",
         "plugins/webida.editor.code-editor.content-assist.css.css-smart",
         "plugins/webida.editor.code-editor.content-assist.css.default",
         "plugins/webida.editor.code-editor.content-assist.html.htmlsmart",
