+++ apps/ide/src/plugins/webida.editor.code-editor.content-assist.comments/plugin.json
@@ -1,0 +1,18 @@
+{
+    "name": "webida.editor.code-editor.content-assist.comments",
+    "description": "Code Editor Contenet assist plugin for comments",
+    "version": "0.1.0",
+    "requirement": "webida.editor.code-editor",    
+    "extensions" : {        
+        "webida.editor.code-editor:contentassist" : [
+            { 
+                "controlModule" : "./CommentsControl", 
+                "engineModule" : "",
+                "langMode" : "*",
+                "engineName" : "Comments",
+                "hinterModes" : [],
+                "hinterNames" : []
+            }
+        ]
+    }
+}
